+++ pkg/gui/filetree/file_change_manager_test.go
@@ -1,4 +1,4 @@
-package gui
+package filetree
 
 import (
 	"testing"
